==> 01_data_curation.ipynb <==
# <p style="text-align: justify;"><div class="alert alert-info" role="alert">Data curation</div></p>

## `Ana González Guerra` 

### ` Student of the master in Data Science at the University of Cantabria` 

## Index<a class="anchor" id="index"></a>
* [Loading data](#1)
* [Filtering missing values](#2)
    * [Physical predictors dataframe](#3)
    * [Genetic predictors dataframe](#4)
    * [Target variables dataframe](#5)
* [Dividing data into more manageable data frames](#6)
* [References](#ref)

## Loading data <a class="anchor" id ="1" ></a>

* [Returning to the index of contents](#index)

import pandas as pd
import numpy as np

We worked from two datasets which shared observations and were merged in order to have a higher number of observations.

### Loading old dataset

handle_old = pd.read_csv('DATOS_TFM_BRUTO.csv', delimiter = ",")

handle_old.head()

### Loading new dataset

handle_new = pd.read_csv('nuevos_datos_27_03_20.csv',delimiter=';',engine='python')
#engine = 'python' gives the possibility to work with csv files with a strange codification that normally python
#does not how to read


handle_new.head()

For joining both files we need a commom key, for the particular case it will be 'CODIGO', thus the column 'código' will be deleted with the aim of avoiding confusions. 

handle_new = handle_new.drop(columns = ['código'])


The new dataset contains a column called 'numero' that it was used for make a count of the individuals as predictors were being measured. It can be eliminated, it does not apport meaning to the analysis.

handle_new = handle_new.drop(columns = ['numero'])

Taking into account that 'PerCad' predictor (hip circumference) , it is a specific predictor for women individuals and the number of women in the sample it is 39, we decided to delete this column and the 'Sexo' predictor (gender of the individual) due to the impossibility of make sub-samples of the original data for man and women with such a low number of observations.

handle_new[handle_new['Sexo'] == 1].shape #  is the coding for a woman individual

handle_new = handle_new.drop(columns = ['PerCad','Sexo'])

handle_old = handle_old.drop(columns = ['Sexo'])

Now, we check if the rows at the end of the datasets are completely occupied with null values:

handle_old.tail()

handle_new.tail()

Besides of rows with zero values along their length we have a row without values for physical predictors and only values for genetic predictors (row 129). It seems reasonable to delete this row plus the last three rows of both datasets (filled with null values).

handle_old = handle_old.iloc[:130,:]

handle_new = handle_new.iloc[:129,:]

After discarding these rows, we merge the data sets into one by selecting the rows present in both data sets, the ones that only the handle_old data set apports, the ones that only the handle_new data set apports, and the rows that both include.

df = handle_old.merge(handle_new, how='outer', indicator='union')#merge

df[df['union'] == 'both'] #selecting all the rows

If we select the rows that are in both datasets we have a total of 102 rows.

df_merged = df[df['union'] == 'both'] 

## Filtering missing values <a class="anchor" id="2"></a>

* [Returning to the index of contents](#index)

In order to work always with the same null value format we will convert all the existing ones to np.nan:

df_merged = df_merged.replace({'No Call':np.nan, 'Invalid':np.nan, '-:-':np.nan})

Now, the null values have been standardised, we analyse their presence in our data with the function missing_values_table(df) [[1]](#koehrsen2018).

def missing_values_table(df):
        # Total missing values
        mis_val = df.isnull().sum()
        
        # Percentage of missing values
        mis_val_percent = 100 * df.isnull().sum() / len(df)
        
        # Make a table with the results
        mis_val_table = pd.concat([mis_val, mis_val_percent], axis=1)
        
        # Rename the columns
        mis_val_table_ren_columns = mis_val_table.rename(
        columns = {0 : 'Missing Values', 1 : '% of Total Values'})
        
        # Sort the table by percentage of missing descending
        mis_val_table_ren_columns = mis_val_table_ren_columns[
            mis_val_table_ren_columns.iloc[:,1] != 0].sort_values(
        '% of Total Values', ascending=False).round(1)
        
        # Print some summary information
        print ("Your selected dataframe has " + str(df.shape[1]) + " columns.\n"      
            "There are " + str(mis_val_table_ren_columns.shape[0]) +
              " columns that have missing values.")
        
        # Return the dataframe with missing information
        return mis_val_table_ren_columns

col_mis_merged = missing_values_table(df_merged)#columns of the dataframe with missing values
col_mis_merged

We have a high number of columns with missing values, but not all the columns have a high number of missing values. Let's look at the columns with more than 5 missing values:

col_mis_merged[col_mis_merged['Missing Values'] >=5]

There is a total of 209 columns with missing values, it can be appreciated that columns with the highest number of missing values are the those referring to the SNPs. We know that some SNPs were replicated due to problems detecting genotypes, these SNPs are: rs1800012, rs3213221 and rs6265. 

Besides of these SNPs replicated, the SNP 'rs3025058' has a 100% of missing values, that does not contain any value, this column is directly removed:

df_merged = df_merged.drop(columns = ['rs3025058'])

Since there are other SNPs with a greater number of missing values, we will determine if there are more replicated SNPs, in addition to the ones mentioned above. To detect these possible replicas we draw from the premise that the columns that are a duplicate will have a name ending in '.1'.

snp = pd.DataFrame()
for i in df_merged:
    if i.endswith('.1'):
        snp = snp.append(df_merged[i])
snp = snp.T
print('Replicated SNPs:',snp.columns)

It seems that the only SNPs replicated are the initial ones mentioned above. With the aim of losing the lesser amount of information for these SNPs we will make a combination of the original and replica column into one column, based on the next criteria (these criteria are using the example of a predictor with possible genotypes: A:A, A:G, G:G):

* If we have A:A in one column, and G:G in the other, the result in the final column is A:G.
* If we have A:G in one column, and G:G in the other, the result in the final column is A:G.
* If we have A:A or G:G in both columns, the result in the final column is A:A or G:G, respectively.
* If we have a null value in one column and an actual value in the other, we take the value of the column with actual value.


Possible genotypes for 'rs1800012':

df_merged['rs1800012'].unique()

#SNP  'rs1800012' and replica 

for i in np.arange(len(df_merged)):
    
    if df_merged['rs1800012'].values[i] is np.nan and df_merged['rs1800012.1'].values[i] is not np.nan:
        df_merged['rs1800012'].values[i] = df_merged['rs1800012.1'].values[i]
    
    elif df_merged['rs1800012'].values[i] is not np.nan and df_merged['rs1800012.1'].values[i] is np.nan:
        pass
    
    elif df_merged['rs1800012'].values[i] == df_merged['rs1800012.1'].values[i]:
        pass
    
    elif df_merged['rs1800012'].values[i] is not df_merged['rs1800012.1'].values[i]:
        if df_merged['rs1800012'].values[i] is not np.nan and df_merged['rs1800012.1'].values[i] is not np.nan:
            df_merged['rs1800012'].values[i] = 'G:T'
    else:
        pass
    

        
        

Possible genotypes for 'rs3213221':

df_merged['rs3213221'].unique()

#SNP 'rs3213221' and replica
for i in np.arange(len(df_merged)):
    
    if df_merged['rs3213221'].values[i] is np.nan and df_merged['rs3213221.1'].values[i] is not np.nan:
        df_merged['rs3213221'].values[i] = df_merged['rs3213221.1'].values[i]
    
    elif df_merged['rs3213221'].values[i] is not np.nan and df_merged['rs3213221.1'].values[i] is np.nan:
        pass
    
    elif df_merged['rs3213221'].values[i] == df_merged['rs3213221.1'].values[i]:
        pass
    
    elif df_merged['rs3213221'].values[i] is not df_merged['rs3213221.1'].values[i]:
        if df_merged['rs3213221'].values[i] is not np.nan and df_merged['rs3213221.1'].values[i] is not np.nan:
            df_merged['rs3213221'].values[i] = 'C:G'
    else:
        pass
    

Possible genotypes for 'rs6265':

df_merged['rs6265'].unique()

#SNP 'rs6265' and replica
for i in np.arange(len(df_merged)):
    
    if df_merged['rs6265'].values[i] is np.nan and df_merged['rs6265.1'].values[i] is not np.nan:
        df_merged['rs6265'].values[i] = df_merged['rs6265.1'].values[i]
    
    elif df_merged['rs6265'].values[i] is not np.nan and df_merged['rs6265.1'].values[i] is np.nan:
        pass
    
    elif df_merged['rs6265'].values[i] == df_merged['rs6265.1'].values[i]:
        pass
    
    elif df_merged['rs6265'].values[i] is not df_merged['rs6265.1'].values[i]:
        if df_merged['rs6265'].values[i] is not np.nan and df_merged['rs6265.1'].values[i] is not np.nan:
            df_merged['rs6265'].values[i] = 'A:G'
    else:
        pass
    

Now, we delete the replica columns:

df_merged = df_merged.drop(columns = ['rs1800012.1','rs3213221.1','rs6265.1'])

After having solved the replicas problem, we are going to look for other SNPs with a high number of missing values, with the aim to decide if they are included or not in the analysis.

missing_values_table(df_merged)

We decided to delete the columns with a percentage equal to or greater than 26.5% in missing values to avoid genetic predictors with few data or observations.

missing_df = missing_values_table(df_merged);
missing_columns = list(missing_df[missing_df['% of Total Values'] >= 26.5].index)
print('Columns removed %d.' % len(missing_columns))

df_merged= df_merged.drop(columns = list(missing_columns))

Once we have eliminated the columns with the greatest number of missing values, we eliminate the rows that have at least one null value among their values. 

df_merged = df_merged.dropna(how='any') 

After this filtering we no longer have missing values:

missing_values_table(df_merged)

We check for the data dimension:

df_merged.shape

Data has acceptable dimensions. From now we will work with this data.

But first, it is necessary to delete the 'CODIGO' column with the aim of deleting the individual identifiers, they are not essential for the modelling and we avoid the risk of compromising their personal identity.

df_merged = df_merged.drop(columns = ['CODIGO'])

Once the datasets have been joined, we no longer need the 'union' column, so we eliminate it.

df_merged= df_merged.drop(columns = ['union'])

We export the final dataset:

df_merged.to_csv('df_merged.csv',sep = ',')

## Dividing data into more manageable data frames <a class="anchor" id="6"></a>

* [Returning to the index of contents](#index)

### Physical predictors dataframe <a class="anchor" id="3"></a>

* [Returning to the index of contents](#index)

To construct the data framework for physical predictors, we exclude all target variables (those starting with "W1M", "W1m", "W2M", "W2m", "W1D", "W2D", "W1-2") and genetic predictors (those starting with "rs").

handle_physical = pd.DataFrame()
for i in df_merged:
    if i.lower().startswith('w1m') or  i.lower().startswith('w2m') or i.lower().startswith('w1d')or i.lower().startswith('w2d')or i.lower().startswith('w1-2'):
        pass
    elif i.startswith('rs'): 
        pass
    else:
        handle_physical = handle_physical.append(df_merged[i])
handle_physical = handle_physical.T
handle_physical

We check handle_physical columns:

handle_physical.columns

We have three kinds of fat percentage but it should be two of them, we look for the error:

handle_physical[['tpcgrasa','tpcgrasa1','tpcgrasa2']]

'tpcgrasa' is the same as 'tpcgrasa1', so we eliminate 'tpcgrasa':

handle_physical = handle_physical.drop(columns = ['tpcgrasa'])

These are the physical predictors that will be considered for the modelling:

handle_physical.columns

print('Number of physical predictors:',len(handle_physical.columns))

We export physical dataframe:

handle_physical.to_csv('handle_physical_fusion_31_03_20.csv', sep=',') 

### Genetic predictors dataframe <a class="anchor" id="4"></a>

* [Returning to the index of contents](#index)

handle_genetic = pd.DataFrame()
for i in df_merged:
    if i.startswith('rs'): 
        handle_genetic = handle_genetic.append(df_merged[i])
    else:
        pass
        
handle_genetic = handle_genetic.T


These will be the genetic predictors considered for the modelling:

print(list(handle_genetic.columns))

print('Number of physical predictors:',len(handle_genetic.columns))

We export handle_genetic dataframe:

handle_genetic.to_csv('handle_genetic_fusion_31_03_20.csv', sep=',') 

### Target variables dataframe <a class="anchor" id="5"></a>

* [Returning to the index of contents](#index)

handle_target = pd.DataFrame()
for i in df_merged:
    if i.lower().startswith('w1m') or  i.lower().startswith('w2m') or i.lower().startswith('w1d')or i.lower().startswith('w2d')or i.lower().startswith('w1-2'):
        handle_target = handle_target.append(df_merged[i])
    else:
        pass
handle_target = handle_target.T

These will be the target variables considered for the modelling:

handle_target.columns

print('Number of target variables:',len(handle_target.columns))

We export target variables dataframe:

handle_target.to_csv('handle_target_fusion_31_03_20.csv',sep = ',')

## References <a class="anchor" id="ref"></a>

* [Returning to the index of contents](#index)

[1] Koehrsen, W. (2018) Missing values. Available at: <a href="https://github.com/WillKoehrsen/machine-learning-project-walkthrough/blob/master/Machine Learning Project Part 1.ipynb">https://github.com/WillKoehrsen/machine-learning-project-walkthrough/blob/master/Machine Learning Project Part 1.ipynb</a> (Accessed: March 31, 2020).
<a class="anchor" id="koehrsen2018"></a>
