==> 02_physical_predictors_transforming_for_multiple_linear_regression.ipynb <==
# <p style="text-align: justify;"><div class="alert alert-info" role="alert">Transforming physical predictors for multiple linear regression</div></p>

## `Ana González Guerra` 

### ` Student of the master in Data Science at the University of Cantabria`  

## Index<a class="anchor" id="index"></a>
* [Loading physical data](#1)
* [Analysis of predictors with zero or close to zero variance](#2)
* [Analysis of linear dependencies between predictors using QR matrix decomposition](#3)
* [Centering and scaling](#4)
* [References](#ref)

## Loading physical data <a class="anchor" id="1"></a>

* [Returning to the index of contents](#index)

handle_physical <- read.csv('handle_physical_fusion_31_03_20.csv',row.name = 1)
head(handle_physical)
names(handle_physical)

In R code the % symbol is not well detected but it still has representation, instead of a % we have 'X.'

In order to avoid interferences with physical predictors relating to FFM1 and similar  (because the intention when we performed the regression model is using W1_2MedDifWRelFFMI as target variable), we remove those physical predictors that have the substring "FM" in their name:

handle_physical <- handle_physical[!grepl('FM',names(handle_physical))] #eliminamos las variables que en su nombre contienen FM
names(handle_physical)

## Analysis of predictors with zero or close to zero variance<a class="anchor" id="2"></a>

* [Returning to the index of contents](#index)

If frequency rate is bigger than a threshold previously established (by default 95/5) and the percent of unique values is smaller than its respectiverly thresholds (by default 10), we could consider that the predictor has a variance near to zero. By default, nearZeroVar function returns the predictors that are considered problematic in relation with the variance. On the other hand, with argument saveMetrics = True we can see the different metrics considered for each one ot the selected predictors as problematic, those that present a TRUE value for nzv column (near zero variance) [[1]](#kuhn2019).

library(caret)

handle_physical_nzv <- nearZeroVar(handle_physical,saveMetrics = TRUE)

handle_physical_nzv[handle_physical_nzv$nzv == TRUE,]

physical_problematicas <- rownames(handle_physical_nzv[handle_physical_nzv$nzv == TRUE,])

It seems physical predictors does not present problems with variance.

## Analysis of linear dependencies between predictors using QR matrix decomposition<a classs="anchor" id="3"></a>

* [Returning to the index of contents](#index)

The findLinearCombos() function uses the QR decomposition of a matrix to list sets of linear combinations (if any). For each linear combination, the number of columns to be removed from the matrix will be increased to check if the dependencies have been resolved. findLinearCombos() will also return a vector with the column positions that can be removed to resolve the situation [[2]](#kuhn2019b).

findLinearCombos matrix 

comboInfo <- findLinearCombos(handle_physical)
comboInfo

comb_lineal <- comboInfo$linearCombos
names(comb_lineal) <- 'cb'

comb_lineal$cb

We can appreciate a linear combinatio for the next physical predictors:

head(handle_physical[comb_lineal$cb])

Where the suggest predictor to remove is 'PliegueMus':

head(handle_fisic[comboInfo$remove])

Since by default the function takes as a predictor to eliminate always the first predictor of the combination, to preserve the biological sense and considering that 'Sumat' is a combination of the rest of fat fold variables (it is the sum of all of them) it seems more reasonable to eliminate 'Sumat' than one of the variables that contributes to the construction of 'Sumat'.

Let's check what happens if we eliminate 'Sumat':

comboInfo_2 <- findLinearCombos(handle_physical[-comb_lineal$cb[2]])
comboInfo_2

Without Sumat we no longer have linear combinations among the physical predictors, so now we elimnate it:

handle_physical <- handle_physical[,-comboInfo$remove]

## Centering and scaling <a class="anchor" id ="4"></a>

* [Returning to the index of contents](#index)

summary(handle_physical)

In this review of the physical predictors data we can see that the ranges in which they move can be very different, therefore it is necessary to scale.

handle_physical <- scale(handle_physical, center = TRUE, scale = TRUE)
summary(handle_physical)

We export the .csv with the data transformed.

write.csv(handle_physical, file="handle_physical_transformed_for_multiple_linear_regression.csv", row.names = F)

## References <a class="anchor" id="ref"></a>

* [Returning to the index of contents](#index)

[1] Kuhn, M. (2019) Zero-and near zero-variance predictors. Available at: https://topepo.github.io/caret/pre-processing.html#zero--and-near-zero-variance-predictors (Accessed: April 20, 2020).<a class="anchor" id="kuhn2019"></a>

[2] Kuhn, M. (2019) Linear dependencies. Available at: https://topepo.github.io/caret/pre-processing.html#linear-dependencies (Accessed: April 20, 2020).<a class="anchor" id="kuhn2019b"></a>
