==> 03_variance_analyses_for_genetic_predictors.ipynb <==
# <p style="text-align: justify;"><div class="alert alert-info" role="alert">Analysis of variance in genetic predictors</div></p>

## `Ana González Guerra` 

### ` Student of the master in Data Science at the University of Cantabria` 

## Index<a class="anchor" id="index"></a>
* [Loading data](#1)
* [Analysis of predictors with zero or close to zero variance](#2)


## Loading data <a class="anchor" id="1"></a>

* [Returning to the index of contents](#index)

handle_genetic <- read.csv('handle_genetic_fusion_31_03_20.csv', row.name = 1)
head(handle_genetic)

## Analysis of predictors with zero or close to zero variance <a class="anchor" id="2"></a>

* [Returning to the index of contents](#index)

library(caret)

handle_genetic_nzv <- nearZeroVar(handle_genetic,saveMetrics = TRUE)

handle_genetic_nzv[handle_genetic_nzv$nzv == TRUE,]

problematic_snp <- nearZeroVar(handle_genetic)

names(handle_genetic[,problematic_snp])

These genetic predictors or SNPs present zero or near zero variance, we eliminate them to avoid interferences later in modelling.

handle_genetic <- handle_genetic[,-problematic_snp]
head(handle_genetic)

We export the new dataset with genetic predictor data:

write.csv(handle_genetic, file="handle_genetic_fusion_28_04_20.csv", row.names = F)

From now on, we will use this dataset for genetic data.
