==> 04_genetic_predictors_codification_1_2_3.ipynb <==
# <p style="text-align: justify;"><div class="alert alert-info" role="alert">Transforming the genetic attributes into numeric values using a codification based on log-additive model</div></p>

## `Ana González Guerra` 

### ` Student of the master in Data Science at the University of Cantabria` 

## Index:<a class="anchor" id="index"></a>
* [Loading data](#1)
* [Codification](#2)
* [Subset of interesting SNPs](#3)

## Loading data

* [Returning to the index of contents](#index)

handle_genetic <-  read.csv('handle_genetic_fusion_28_04_20.csv')

head(handle_genetic)

## Codification <a class="anchor" id="2"></a>

* [Returning to the index of contents](#index)

head(handle_genetic)

for (i in 1: ncol(handle_genetic)){
    handle_genetic[,i] <- as.numeric(handle_genetic[,i])
}

Thus the levels for each genetic variable will be 1 for the first level in alphabetical order (within the possibilities of the particular variable), 2 for the second one, and 3 for the third one.

head(handle_genetic)

We check that we have numbers and no strings

class(handle_genetic[,1])

We export the csv with the coded genetic data

write.csv(handle_genetic, file="handle_genetic_cod_28_04_20.csv", row.names = F)

## Subset of interesting SNPs<a class="anchor" id="3"></a>

* [Returning to the index of contents](#index)

handle_genetic_top <- handle_genetic[,c('rs1815739',
 'rs4343',
 'rs1049434',
 'rs1042713',
 'rs699',
 'rs1800795',
 'rs4253778',
 'rs4680',
 'rs8111989',
 'rs35767',
 'rs11549465',
 'rs2070744')]

head(handle_genetic_top)

write.csv(handle_genetic_top, file="handle_genetic_top_cod_28_04_20.csv", row.names = F)
