==> 08_physical_predictors_with_sumat_regression.ipynb <==
# <p style="text-align: justify;"><div class="alert alert-info" role="alert">Multiple Linear Regression with physical predictors and Sumat as unique fat fold predictor</div></p>

## `Ana González Guerra` 

### ` Student of the master in Data Science at the University of Cantabria` 

## Index<a class="anchor" id="index"></a>
* [Loading data](#1)
    * [Physical predictors](#2)
    * [Target variables](#3)
* [Train/test partition](#4)
* [Multiple Linear Regression](#5)
   

## Loading data <a class="anchor" id ="1" ></a>

* [Returning to the index of contents](#index)

### Physical predictors <a class="anchor" id="2"></a>

* [Returning to the index of contents](#index)

handle_physical <- read.csv('handle_physical_transformed_for_multiple_linear_regression.csv')
head(handle_physical)
names(handle_physical)

To avoid interference with the predictors related to FFM1 and similars, we eliminated the predictors that in their name contain the FM substring.

handle_physical <- handle_physical[!grepl('FM',names(handle_physical))] 

We delete all the fat folds predictors that are not Sumat:

handle_physical <- handle_physical[!grepl('Pliegue',names(handle_physical))] 

head(handle_physical)

### Target variables <a class="anchor" id="3"></a>

* [Returning to the index of contents](#index)

handle_target <- read.csv('handle_target_fusion_31_03_20.csv', row.name = 1)

head(handle_target)

Seleccionamos la variable objetivo W1-2MedDifWRelFFMI:

target <- handle_target['W1.2MedDifWRelFFM1']
head(target)

We create the dataframe for the analysis:

data_physical <- cbind(target, handle_physical)

## Train/test partition <a class="anchor" id ="4"></a>

* [Returning to the index of contents](#index)

#### Partición train/test <a class="anchor" id="18" ></a>

library(caret)

set.seed(10)
id_df <- createDataPartition(data_physical$W1.2MedDifWRelFFM1,p = 0.85,list = FALSE)
data_physical_train <- data_physical[id_df,]
data_physical_test <- data_physical[-id_df,]

dim(data_physical_train)

dim(data_physical_test)

## Multiple Linear Regression <a class="anchor" id ="5"></a>

* [Returning to the index of contents](#index)

Training:

model_physical_train <- lm(W1.2MedDifWRelFFM1~., data_physical_train)

summary(model_physical_train)

rmse <- function(obs, est) {
  sqrt(mean((obs - est)^2))
}#with this function we will compute the prediction error in train and test

Train error:

rmse(data_physical_train[,1], model_physical_train$fitted.values)

Test error:

pred_test <- predict(model_physical_train, newdata = data_physical_test)

rmse(data_physical_test[,1], pred_test)
